day 9 - I had to change all longs to int -.-

+++ src/test/kotlin/com/github/shmvanhouten/adventofcode2019/day02/IntCodesTest.kt
@@ -20,59 +20,59 @@ class IntCodesTest {
         @Test
         internal fun `1, 0, 0, 0 becomes 2, 0, 0, 0`() {
             val intCode = IntCode(1, 0, 0, 0, 99)
-            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(2, 0, 0, 0, 99)))
+            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(2L, 0L, 0L, 0L, 99)))
         }
 
         @Test
         internal fun `1, 1, 1, 0 becomes 2, 1, 1, 0`() {
             val intCode = IntCode(1, 1, 1, 0, 99)
-            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(2, 1, 1, 0, 99)))
+            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(2L, 1L, 1L, 0L, 99)))
         }
 
         @Test
         internal fun `1, 0, 0, 1 becomes 1, 2, 0, 1`() {
             // intCode[3] determines the position of the new value
             val intCode = IntCode(1, 0, 0, 1, 99)
-            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(1, 2, 0, 1, 99)))
+            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(1L, 2L, 0L, 1L, 99)))
         }
 
         @Test
         internal fun `1, 2, 1, 1 becomes 1, 4, 1, 1`() {
             val intCode = IntCode(1, 2, 1, 1, 99)
-            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(1, 3, 1, 1, 99)))
+            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(1L, 3L, 1L, 1L, 99)))
         }
 
         @Test
         internal fun `1,0,0,0,99 becomes 2,0,0,0,99`() {
             // one of the examples
             val intCode = IntCode(1, 0, 0, 0, 99, 99)
-            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(2, 0, 0, 0, 99, 99)))
+            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(2L, 0L, 0L, 0L, 99, 99)))
         }
 
         @Test
         internal fun `2,0,0,1 becomes 2, 4, 0, 1`() {
             val intCode = IntCode(2, 0, 0, 1, 99)
-            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(2, 4, 0, 1, 99)))
+            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(2L, 4L, 0L, 1L, 99)))
         }
 
         //2,3,0,3,99 becomes 2,3,0,6,99 (3 * 2 = 6).
         @Test
         internal fun `2,3,0,3,99 becomes 2,3,0,6,99`() {
             val intCode = IntCode(2, 4, 4, 5, 99, 0)
-            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(2, 4, 4, 5, 99, 9801)))
+            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(2L, 4L, 4L, 5L, 99, 9801)))
         }
 
         //2,4,4,5,99,0 becomes 2,4,4,5,99,9801 (99 * 99 = 9801).
         @Test
         internal fun `2,4,4,5,99,0 becomes 2,4,4,5,99,9801`() {
             val intCode = IntCode(2, 4, 4, 5, 99, 0)
-            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(2, 4, 4, 5, 99, 9801)))
+            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(2L, 4L, 4L, 5L, 99, 9801)))
         }
 
         @Test
         internal fun `99 stops the program`() {
             val intCode = IntCode(99, 0, 0, 0)
-            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(99, 0, 0, 0)))
+            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(99, 0L, 0L, 0L)))
         }
 
         @Test
@@ -85,12 +85,12 @@ class IntCodesTest {
         @Test
         internal fun `we run the next block when this block is finished`() {
             assertThat(
-                Computer(intCode = IntCode(1, 0, 0, 1, 1, 0, 0, 1, 99)).run(),
-                equalTo(listOf(1, 2, 0, 1, 1, 0, 0, 1, 99))
+                Computer(intCode = IntCode(1L, 0L, 0L, 1L, 1L, 0L, 0L, 1L, 99)).run(),
+                equalTo(listOf(1L, 2L, 0L, 1L, 1L, 0L, 0L, 1L, 99))
             )
             assertThat(
-                Computer(intCode = IntCode(1, 0, 0, 1, 1, 0, 0, 4, 99)).run(),
-                equalTo(listOf(1, 2, 0, 1, 2, 0, 0, 4, 99))
+                Computer(intCode = IntCode(1L, 0L, 0L, 1L, 1L, 0L, 0L, 4L, 99)).run(),
+                equalTo(listOf(1L, 2L, 0L, 1L, 2L, 0L, 0L, 4L, 99))
             )
         }
 
@@ -98,7 +98,7 @@ class IntCodesTest {
         @Test
         internal fun `1,1,1,4,99,5,6,0,99 becomes 30,1,1,4,2,5,6,0,99`() {
             val intCode = IntCode(1,1,1,4,99,5,6,0,99)
-            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(30,1,1,4,2,5,6,0,99)))
+            assertThat(Computer(intCode = intCode).run(), equalTo(listOf(30L,1L,1L,4L,2L,5L,6L,0L,99)))
         }
 
         @Test
@@ -118,6 +118,6 @@ class IntCodesTest {
         }
     }
 
-    private fun parseInput(input: String) = input.split(',').map { it.toInt() }
+    private fun parseInput(input: String) = input.split(',').map { it.toInt().toLong() }
 
 }
